==> 01_EDHD_Ch9_Dataset_ZIP_extraction.ipynb <==
# # To recursively remove the data directory!!!


# import shutil
# import os

# folder_path = "../Processed_Data"  # change to your folder name

# if os.path.exists(folder_path):
#     shutil.rmtree(folder_path)
#     print(f"Folder '{folder_path}' and all its contents have been removed.")
# else:
#     print(f"Folder '{folder_path}' does not exist.")


# To extract the zipped dataset!!!

import zipfile
import os

zip_path = "../EDHD_Ch9_Dataset.zip"
extract_dir = "../Extracted_Dataset"

if not os.path.exists(zip_path):
    print(f"Zip file '{zip_path}' not found mate.. Please provide it in the current directory.")
else:
    if os.path.exists(extract_dir):
        print("Mate dataset is extracted from zip file already")
    else:
        with zipfile.ZipFile(zip_path, 'r') as zip_ref:
            zip_ref.extractall(extract_dir)
        print(f"Extracted to: {extract_dir}")

    # List folder structure
    print("\nFolder structure:\n")
    for root, dirs, files in os.walk(extract_dir):
        level = root.replace(extract_dir, "").count(os.sep)
        indent = " " * 4 * level
        print(f"{indent}{os.path.basename(root)}/")
        subindent = " " * 4 * (level + 1)
        for f in files:
            print(f"{subindent}{f}")
